arbitrage_bot/requirements.txt
# Core
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.6
greenlet>=3.0.3,<3.3
idna==3.10
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.1.1
python-telegram-bot==13.15
pytz==2025.2
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.43
tornado==6.1
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
python-dateutil==2.9.0.post0

# For async_base_scraper
httpx==0.27.0
beautifulsoup4==4.12.3
playwright==1.47.0
celery==5.4.0

# For orchestrator (cache/queue)
redis==5.0.8

# For BaseScraper (browser fallback)
selenium==4.25.0
chromedriver-autoinstaller==0.6.4
lxml==5.3.0
